+++ CHANGELOG.md
@@ -33,3 +33,5 @@
 
 #### 2014.08.06
 - blog-roll images circle
+- Revert circle image for blog-roll
+- If no blog-roll image exists, remove image from layout